backend312/pyproject.toml
[tool.poetry]
name = "project-finscrape"
version = "0.1.0"
description = "AI to turn websites into summaries based on provided criteria."
authors = ["Adam Huang"]

[tool.poetry.dependencies]
# python = ">=3.10.0,<3.12"
python = ">=3.11.0, <=3.12.1"
django = "^5.0.6"
requests = "^2.32.2"
bs4 = "^0.0.2"
html5lib = "^1.1"
openai = "^1.30.3"
python-dotenv = "^1.0.1"
asyncio = "^3.4.3"
nltk = "^3.8.1"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"
psycopg2-binary = "^2.9.9"
pypiwin32 = "^223"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
# Python code for type errors
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"